refactor(api): Share utils module
