+++ wagtail/api/v1/serializers.py
@@ -13,7 +13,7 @@ from rest_framework import relations
 from wagtail.utils.compat import get_related_model
 from wagtail.wagtailcore import fields as wagtailcore_fields
 
-from .utils import get_full_url, pages_for_site
+from ..shared.utils import get_full_url, pages_for_site
 
 
 def get_object_detail_url(context, model, pk):
