+++ wagtail/api/v2/filters.py
@@ -7,7 +7,7 @@ from taggit.managers import _TaggableManager
 from wagtail.wagtailcore.models import Page
 from wagtail.wagtailsearch.backends import get_search_backend
 
-from .utils import BadRequestError, pages_for_site
+from ..shared.utils import BadRequestError, pages_for_site
 
 
 class FieldsFilter(BaseFilterBackend):
